Extract reusable methods to test/lib/utils
