+++ balancer-js/src/modules/zaps/bbausd2-migrations/bbausd1.integration.spec.ts
@@ -13,16 +13,11 @@ import {
 import { BigNumber, parseFixed } from '@ethersproject/bignumber';
 import { Contracts } from '@/modules/contracts/contracts.module';
 import { ADDRESSES } from './addresses';
-import { setBalance } from '@nomicfoundation/hardhat-network-helpers';
-import { parseEther } from '@ethersproject/units';
 import { JsonRpcSigner } from '@ethersproject/providers';
 import { MaxUint256 } from '@ethersproject/constants';
 import { Migrations } from '../migrations';
-
-import { Interface } from '@ethersproject/abi';
 import { PoolsProvider } from '@/modules/pools/provider';
-const liquidityGaugeAbi = ['function deposit(uint value) payable'];
-const liquidityGauge = new Interface(liquidityGaugeAbi);
+import { getErc20Balance, move, stake } from '@/test/lib/utils';
 
 /*
  * Testing on GOERLI
@@ -58,19 +53,6 @@ const migrations = new Migrations(network);
 
 const holderAddress = '0xd86a11b0c859c18bfc1b4acd072c5afe57e79438';
 const relayer = addresses.relayer;
-
-const getErc20Balance = (token: string, holder: string): Promise<BigNumber> =>
-  contracts.ERC20(token, provider).balanceOf(holder);
-
-// https://hardhat.org/hardhat-network/docs/guides/forking-other-networks#impersonating-accounts
-// WARNING: don't use hardhat SignerWithAddress to sendTransactions!!
-// It's not working and we didn't have time to figure out why.
-// Use JsonRpcSigner instead
-const impersonateAccount = async (account: string) => {
-  await provider.send('hardhat_impersonateAccount', [account]);
-  await setBalance(account, parseEther('10000'));
-  return provider.getSigner(account);
-};
 
 const signRelayerApproval = async (
   relayerAddress: string,
@@ -109,37 +91,6 @@ const reset = async () =>
     },
   ]);
 
-const move = async (
-  token: string,
-  from: string,
-  to: string
-): Promise<BigNumber> => {
-  const holder = await impersonateAccount(from);
-  const balance = await getErc20Balance(token, from);
-  await contracts.ERC20(token, provider).connect(holder).transfer(to, balance);
-
-  return balance;
-};
-
-const stake = async (
-  signer: JsonRpcSigner,
-  balance: BigNumber
-): Promise<void> => {
-  await (
-    await contracts
-      .ERC20(fromPool.address, provider)
-      .connect(signer)
-      .approve(fromPool.gauge, MaxUint256)
-  ).wait();
-
-  await (
-    await signer.sendTransaction({
-      to: fromPool.gauge,
-      data: liquidityGauge.encodeFunctionData('deposit', [balance]),
-    })
-  ).wait();
-};
-
 describe('bbausd migration execution', async () => {
   let signer: JsonRpcSigner;
   let signerAddress: string;
@@ -157,7 +108,12 @@ describe('bbausd migration execution', async () => {
     authorisation = await signRelayerApproval(relayer, signerAddress, signer);
     // Transfer tokens from existing user account to signer
     // We need that to test signatures, because hardhat doesn't have impersonated accounts private keys
-    balance = await move(fromPool.address, holderAddress, signerAddress);
+    balance = await move(
+      fromPool.address,
+      holderAddress,
+      signerAddress,
+      provider
+    );
 
     const config = {
       network,
@@ -183,8 +139,8 @@ describe('bbausd migration execution', async () => {
     const addressOut = staked ? toPool.gauge : toPool.address;
     // Store balance before migration
     const before = {
-      from: await getErc20Balance(addressIn, signerAddress),
-      to: await getErc20Balance(addressOut, signerAddress),
+      from: await getErc20Balance(addressIn, provider, signerAddress),
+      to: await getErc20Balance(addressOut, provider, signerAddress),
     };
 
     const amount = before.from;
@@ -231,8 +187,8 @@ describe('bbausd migration execution', async () => {
     console.log('Gas used', receipt.gasUsed.toString());
 
     const after = {
-      from: await getErc20Balance(addressIn, signerAddress),
-      to: await getErc20Balance(addressOut, signerAddress),
+      from: await getErc20Balance(addressIn, provider, signerAddress),
+      to: await getErc20Balance(addressOut, provider, signerAddress),
     };
 
     console.log(expectedBpts);
@@ -249,7 +205,7 @@ describe('bbausd migration execution', async () => {
     beforeEach(async function () {
       this.timeout(20000);
       // Stake them
-      await stake(signer, balance);
+      await stake(signer, fromPool.address, fromPool.gauge, balance);
     });
 
     it('should transfer tokens from stable to boosted - using exact bbausd2AmountOut from static call', async () => {
